File structure reorganization
- split modules into input, output and processing nodes to reduce file length
- add video encoding node
