+++ .pylintrc
@@ -13,3 +13,4 @@ disable =
     R0201, # problems with abstract methods
     W0511, # allow TODO/FIXME for now (should be removed when going stable)
     E1136, # subscriptable error when intializing object with None first
+    R1732, # problems with file opening
